fix: start a discussion button shouldn't be both link and button
For accessibility and also functionality it's better to choose either a link or a button. As this action is to navigate to a new page it makes sense to have this as a link that is styled as a button. This also allows us to style this a bit nicer :sparkles: