feat: primitive support chat page
STILL LOTS TO DO ON THIS!
This is just an initial pass at this to have something to grow on. There is actually a lot to do here still, styling aside.
Notably all the permissions stuff and grabbing the actual correct threads, permissions etc. and also having the threadID in the url for thread details. Coming coming