refactor: support pages should have threadId in url
Currently these support pages only have the selected thread in memory rather than in the url. This refactor uses next's layout and page feature to have the support chat as well as the thread detail drawer depending on the url.

Still lots of styling to be done! This just gets a base out there to build on :rocket: