fix: broken deployment & support page navigation
Didn't realise that you can't pass children to page.tsx! So have rejigged again so that I'm actually using Next as it should be. Ie. children are only available in layout.tsx.

Also updated the navigation on thread click to be root (starts with `/`) instead of appending the whole url again