Create Button component
We've got buttons that don't look the same throughout the app. Plus some are missing some important accessibility settings (aria-labels, colour contrast, pseudo states etc), this PR creates a Button component that can be used as either a link or a button which should cover the basics of what is required for either of these.

It styles links and buttons to look the same (as we're trying to do that currently for each instance)