style: outlet height should not include header
Currently the outlet height is set to 100vh, which doesn't include the header, so scrolling isn't behaving as expected as the outlet has a height that overflows the viewport height.

This commit takes away the height of the header (just by looking at how much it takes up :eyes:) and sets that to the height of the outlet.

This should stop janky scrolling