style: make support page look half decent
Had to not use Thread or ThreadList cord components as they didn't have the configurability we want. ie. sort direction & styling was quite painful as we're using css modules, so a lot of the styling had to go in global.css which isn't very nice.

So building the chat from scratch, we are using CordMessage though :raised_hands: & has a "new start" text for when the very beginning of the chat - super simple, can review UI at a later date.

This commit doesn't include mobile responsiveness, this is already a fair chunk of code touching a fair few files, so will leave responsiveness to a dedicated commit, especially as the rest of the app isn't responsive yet either.

These changes are all cosmetic, I haven't hooked up to the actual customer's chat yet.