feat: use Customer name and ID for support nav and chat display
This commit uses the getCustomerInfo to grab the relevant data for the currently signed in customer. It's v1 so it currently doesn't account for Cord admins that will have multiple support chats. This will have to be a separate call to grab all the groups separately. (I've added a comment where we put that data)

I've also added 2 buttons, upgrade cord and join cord which will show under support chat in the Sidebar depending on whether the user has a cord account or has supportEnabled.

On the chat itself it renders a very basic text that says you don't have a cord account - this needs to be done in a follow up