Create not logged in page for post details
This was broken because our post details page uses Cord client side hooks which requires a clientAuthToken. For a non-logged in user ththese calls return only the initialised data (ie. none) and so can't render anything.

This kills 2 birds with one stone, in that this page can now be SSR-ed and allow non logged in users to view the page.

Currently our Cord components don't allow us to use SSR, but also they use the hooks under the hood, which means even if they were able to render on the server, they would fail once hydrated as there's no client auth token.

Looks like there is work in flow to not only allow SSR for non data reliant components (ie. Timestamp) but also that we will soon have the option to pass data to our components instead of having them fetched under the hood. This is why I've chosen to implement very rudementary avatar and timestamp, and haven't even bothered with the Reactions yet :eyes:

Proper styling coming in a future PR. I want to get the styling looking better for the client side, then can mimick for ss :)