style: Post Details to look nicer
Really just UI styling of the Post Details pages (both server and client side).
The client side (Cord) components are really painful to style when using css modules :( all the styling has to live in global.css and then we have to match it up with our custom components.

Not really sure if there's a better solution to this - will talk to Ludo / Alberto