Post details should render even if theres no slug in the url
We have the slug of the thread name in the url, however, we aren't actually using that anywhere for fetching data etc. It's just for making SEO love us even more. Because of this, it makes sense for the route with `/post/[postID]` (no slug) to work as well.

Includes some jigging around, putting ServerPost into it's own file.