Add cord stylesheet to the app
Currently we have a little bit of flicker in some components that are being rendered on the Server because they don't have access to the stylesheet that is provided in the SDK.

This commit adds the stylesheets to the html headers (in root layout) so that they are available. We're unsure if this is the way Next wants us to do it, but they also aren't very forthcoming with how it should be done. This works, so for now I'm happy with it :shrug:

There are two stylesheets as one is for onion components (cord-react) and the other is for no-onion components (sdk.latest). Alberto and Ludo are your go-tos if you're struggling with this!