Get ThreadID from our KV store
Use our kv store and increment the nextThreadID value.  This way our posts will increment sequentially and be much more human readable.

Importantly we need to make sure that vercel does not cache fetches within the store.  Added the directive at the top of the page to force it hit the server rather than the cache.