Simplify our login
Rather than trying to do a silent-login on basically every page load and using cookies to track whether we are attempting and where to return to, let's just not.  This was causing issues with search engines because they don't set cookies so it would just end up in a redirect loop.

Instead, let's just let the session object from Auth0 do it's thing.  If you are logged in, we will pick it up.  If you are logged in to console, clicking sign in will do a silent login.

To still allow the silent login attempt when linking from console, i've added a silent_login=true search param that you can use to cause us to do the silent login.