Fix up a bunch of issues with the left nav
The left nav wasn't obeying a huge number of best practices for accessible
pages. It was mostly to do with an `all: unset` CSS declaration that was turning
off all sorts of very good and useful things. We don't use `all: unset` unless
we know exactly what it's taking away.

Anyway, this change puts hover behaviour back into the navigation properly. It
also makes the `a` tags in the nav full width via `display: block` rather than
`width: 100%`.

This also fixes the minimum height of the left nav to be >= 48px on mobile. This
is important for touch interfaces because of clumsy human fingers. Mobile
navigability impacts SEO. Incidentally, this change moves us from 86 on the
Lighthouse SEO score to 90.