Experiemental - Adding a mark as answer to options menu
This was me testing whether we could add
a mark as answer button to the message options menu
so I could generate some  QandAPage structred data.

Had to change the thread to an experiemental one,
and for annoyed that I had to store the messageID
in a state somewhere, and also the threadID
(thus all the new components with the contexts!)

Also don't show the resolved menu item button when
you are not an admin.

Hoping Cord 4.0 exposes the threadIDs and MessageIDs
within the replacement components 🙏 but this currently
works. When you mark as answer it just updates
the thread metadata with answerMessageID as the message ID
that was chosen.

Test Plan:
Tested locally, then went to console and found
the thread and saw metadata was updated

Reviewers: dmmiller

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/community/pull/3