Dynamically generate title tags from params
Changed this for category and for the posts.

Trying to think whether it's worth generating the
meta description for the posts too...

Test Plan:
Clicked on each category can see the
title changes to  Components | Cord Community, and
also checked the posts.

Don't know if we need to append the | Cord Community
as some site says the domain is automatically added
but some examples show this.

Reviewers: dmmiller

Pull Request: https://github.com/getcord/community/pull/4