Nuke the table layout
I'm not sure why we picked a table layout for the main content list. Probably
because that's what Discourse do? Anyway, it was wreaking havoc on mobile
layouts. What I'm doing here is replacing the table layout with a semantic HTML
layout. I'm using article, blockquote, and time elements to help make it clear
to search engines what is what here.

This also handles narrower mobile layouts much more sensibly. I max-width'd the
content boxes to be 600px. This is for readability. On a huge screen, these
article texts and titles where massively wide, which was really hard to read.