Add solution label in post list and for answer message
When looking at the List of posts, would like to
know if there has been a solution i.e. when we
mark it as a solution.

Did something hacky to get it to render in the client
side thread (i.e. replacing the message timestamp with
timestamp and solution label, CSS specificity issues)