Create a custom TextEditor so that we can swallow Enter as a Submit
This works exactly the same as the built in TextEditor except that when you hit Enter, it sets the shiftKey to be true so that it seems like the user has hit shift+Enter.  That means when the TextEditor processes the key press, it won't tell the Composer to Submit.  It will actually treat it more like a new line.  This behaviour makes more sense for this style of interaction, more like a board where you can post long form content.