Position the composer at the bottom always
Trying to achieve this by using position: fixed on the post container would
mean we'd have to account for the collapsed/expanded state of the sidebar, hence
opted for this solution instead.

This way, we can use the flex + some CSS calculations to figure the position
of the thread container. I used 48px for header container as that never
changes since the avatar is always set to 32px, and the padding to 8px
at all times.

Test Plan:
- Composer is at the bottom of page in /post/id, the thread content remains
scrollable on both mobile and desktop.

Reviewers: 

Pull Request: https://github.com/getcord/community/pull/19