Scroll to the first message (question) automatically
Wee've got a fix in #20 that updates the prop from never to 'auto' once
the viewer has scrolled to the last message, but this doesn't work
yet, and onion team is looking at that.

For now, we can have this fix in.

Test plan:
We see the oldest message first when we land on post page