Revert fix for making nav fixed
Summary:
Having an overflow auto on the content, meant that to scroll in
the posts page we had to click on the content container.
This also means that scrolling in support pages is wonky again, and
the post page no longer has the composer fixed at the bottom -  I'll
add a fix for that separately in a follow up.

Test plan:
Using spacebar to scroll now works without having to first click on
the posts container.