Replace MenuButton rather than Menu
Summary:

The Nonion structure of this component is currently not great.
We have a `MenuButton`, which can decide to not render itself if it has no `menuItems`. The `MenuButton` renders the button with "..." that if you click, opens up a menu.

In Community, we were replacing `MenuButton`'s child, `Menu`, by passing two extra `menuItem` (delete post, mark as solution). However, we were not passing these two items to `MenuButton`, which unfortunately meant it didn't render anything. (because hey, there are no `menuItems`/actions you can perform on other people's messages).

One solution we can do in community is to replace the top level `MenuButton`, so that `menuItems` we add are correctly propagated to `Menu`.

Test plan:
Load Community locally, I can now see the custom buttons.